pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "options-trader"
version = "0.2.0"
description = "Synthetic long strategy scanner"
readme = "README.md"
authors = [{name = "Abhinav Mangat"}]
requires-python = ">=3.9"
dependencies = [
  "yfinance>=0.2.36",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
full = [
  "yfinance>=0.2.36",
  "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/abmangat/Options"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
